requirements.txt
groq
fastapi
uvicorn
openai
langchain
faiss-cpu
pdfplumber
python-docx
python-dotenv
tqdm
scikit-learn
sentence-transformers